feat: Add optional secondary param if query args arne't required
