+++ .changeset/fresh-cups-speak.md
@@ -1,0 +1,7 @@
+---
+'@ts-rest/core': minor
+'@ts-rest/react-query': minor
+'@ts-rest/solid-query': minor
+---
+
+Added the ability to omit the second parameter if there no required query parameters
