--- libs/ts-rest/react-query/src/lib/test-fixtures.ts
@@ -1,107 +1,0 @@
-import { initContract } from '@ts-rest/core';
-import { z } from 'zod';
-
-const c = initContract();
-
-export type Post = {
-  id: string;
-  title: string;
-  description: string | null;
-  content: string | null;
-  published: boolean;
-  authorId: string;
-};
-
-export type User = {
-  id: string;
-  email: string;
-  name: string | null;
-};
-
-const postsRouter = c.router({
-  getPost: {
-    method: 'GET',
-    path: `/posts/:id`,
-    responses: {
-      200: c.response<Post | null>(),
-    },
-  },
-  getPosts: {
-    method: 'GET',
-    path: '/posts',
-    responses: {
-      200: c.response<Post[]>(),
-    },
-    query: z.object({
-      take: z.number().optional(),
-      skip: z.number().optional(),
-    }),
-  },
-  createPost: {
-    method: 'POST',
-    path: '/posts',
-    responses: {
-      200: c.response<Post>(),
-    },
-    body: z.object({
-      title: z.string(),
-      content: z.string(),
-      published: z.boolean().optional(),
-      description: z.string().optional(),
-      authorId: z.string(),
-    }),
-  },
-  mutationWithQuery: {
-    method: 'POST',
-    path: '/posts',
-    responses: {
-      200: c.response<Post>(),
-    },
-    body: z.object({}),
-    query: z.object({
-      test: z.string(),
-    }),
-  },
-  updatePost: {
-    method: 'PUT',
-    path: `/posts/:id`,
-    responses: {
-      200: c.response<Post>(),
-    },
-    body: z.object({
-      title: z.string(),
-      content: z.string(),
-      published: z.boolean().optional(),
-      description: z.string().optional(),
-      authorId: z.string(),
-    }),
-  },
-  patchPost: {
-    method: 'PATCH',
-    path: `/posts/:id`,
-    responses: {
-      200: c.response<Post>(),
-    },
-    body: null,
-  },
-  deletePost: {
-    method: 'DELETE',
-    path: `/posts/:id`,
-    responses: {
-      200: c.response<boolean>(),
-    },
-    body: null,
-  },
-});
-
-// Three endpoints, two for posts, and one for health
-export const router = c.router({
-  posts: postsRouter,
-  health: {
-    method: 'GET',
-    path: '/health',
-    responses: {
-      200: c.response<{ message: string }>(),
-    },
-  },
-});
