+++ package.json
@@ -183,7 +183,7 @@
     "ts-jest": "28.0.8",
     "ts-node": "^10.9.1",
     "tsd": "^0.22.0",
-    "typescript": "4.8.4",
+    "typescript": "4.9.4",
     "url-loader": "^4.1.1",
     "vite": "^3.1.2",
     "vite-plugin-solid": "2.3.0",
